Changed double type for delta into float type, added first test (moving)

+++ Mazzi/FlowerForceSolution/FlowerForceZombie/IMovingEntity.cs
@@ -5,6 +5,6 @@ namespace FlowerForceZombie
     {
         void Move();
 
-        double Delta { get; }
+        float Delta { get; }
     }
 }
